TST: Skip appropriate tests when no PyProj present
This allows us to run the tests without installing PyProj. It adds a
decorator for test functions/fixtures that will skip if importing pyproj
fails.
