+++ tests/calc/test_kinematics.py
@@ -18,7 +18,7 @@ from metpy.calc import (absolute_vorticity, advection, ageostrophic_wind, coriol
                         total_deformation, vorticity, wind_components)
 from metpy.constants import g, omega, Re
 from metpy.testing import (assert_almost_equal, assert_array_almost_equal, assert_array_equal,
-                           get_test_data, needs_cartopy)
+                           get_test_data, needs_cartopy, needs_pyproj)
 from metpy.units import concatenate, units
 
 
@@ -427,6 +427,7 @@ def test_absolute_vorticity_asym():
 
 
 @pytest.fixture
+@needs_pyproj
 def pv_data():
     """Test data for all PV testing."""
     u = np.array([[[100, 90, 80, 70],
@@ -542,6 +543,7 @@ def test_potential_vorticity_baroclinic_wrong_number_of_levels_axis_0(pv_data):
                                        pressure[:1, :, :])
 
 
+@needs_pyproj
 def test_potential_vorticity_baroclinic_isentropic_real_data():
     """Test potential vorticity calculation with real isentropic data."""
     isentlevs = [328, 330, 332] * units.K
@@ -636,6 +638,7 @@ def test_potential_vorticity_baroclinic_isentropic_real_data():
     assert_almost_equal(pvor, true_pv, 14)
 
 
+@needs_pyproj
 def test_potential_vorticity_baroclinic_isobaric_real_data():
     """Test potential vorticity calculation with real isentropic data."""
     pres = [20000., 25000., 30000.] * units.Pa
@@ -744,23 +747,7 @@ def test_potential_vorticity_barotropic(pv_data):
     assert_almost_equal(pv, truth, 10)
 
 
-def test_lat_lon_grid_deltas_geod_kwargs():
-    """Test that geod kwargs are overridden by users #774."""
-    lat = np.arange(40, 50, 2.5)
-    lon = np.arange(-100, -90, 2.5)
-    dx, dy = lat_lon_grid_deltas(lon, lat, a=4370997)
-    dx_truth = np.array([[146095.76101984, 146095.76101984, 146095.76101984],
-                         [140608.9751528, 140608.9751528, 140608.9751528],
-                         [134854.56713287, 134854.56713287, 134854.56713287],
-                         [128843.49645823, 128843.49645823, 128843.49645823]]) * units.meter
-    dy_truth = np.array([[190720.72311199, 190720.72311199, 190720.72311199, 190720.72311199],
-                         [190720.72311199, 190720.72311199, 190720.72311199, 190720.72311199],
-                         [190720.72311199, 190720.72311199, 190720.72311199,
-                          190720.72311199]]) * units.meter
-    assert_almost_equal(dx, dx_truth, 4)
-    assert_almost_equal(dy, dy_truth, 4)
-
-
+@needs_pyproj
 def test_inertial_advective_wind_diffluent():
     """Test inertial advective wind with a diffluent flow."""
     lats = np.array([[50., 50., 50., 50., 50., 50., 50., 50., 50., 50., 50.],
@@ -952,6 +939,7 @@ def test_inertial_advective_wind_diffluent():
 
 
 @pytest.fixture
+@needs_pyproj
 def q_vector_data():
     """Define data for use in Q-vector tests."""
     speed = np.ones((4, 4)) * 50. * units('knots')
