+++ tests/plots/test_declarative.py
@@ -20,7 +20,7 @@ from metpy.io.metar import parse_metar_file
 from metpy.plots import (BarbPlot, ContourPlot, FilledContourPlot, ImagePlot, MapPanel,
                          PanelContainer, PlotObs)
 # Fixtures to make sure we have the right backend
-from metpy.testing import needs_cartopy, set_agg_backend  # noqa: F401, I202
+from metpy.testing import needs_cartopy, needs_pyproj, set_agg_backend  # noqa: F401, I202
 from metpy.units import units
 
 
@@ -28,6 +28,7 @@ MPL_VERSION = matplotlib.__version__[:3]
 
 
 @pytest.mark.mpl_image_compare(remove_text=True, tolerance=0.005)
+@needs_pyproj
 def test_declarative_image():
     """Test making an image plot."""
     data = xr.open_dataset(GiniFile(get_test_data('NHEM-MULTICOMP_1km_IR_20151208_2100.gini')))
@@ -322,6 +323,7 @@ def test_colorfill_no_colorbar(cfeature):
 
 
 @pytest.mark.mpl_image_compare(remove_text=True, tolerance=1.23)
+@needs_pyproj
 def test_global():
     """Test that we can set global extent."""
     data = xr.open_dataset(GiniFile(get_test_data('NHEM-MULTICOMP_1km_IR_20151208_2100.gini')))
